template/pyproject.toml
[tool.poetry]
name = "__APP_NAME__"
version = "0.1.0"
description = "FastAPI project generated by fastapi_starter"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [ { include = "app" } ]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "0.115.0"
"uvicorn" = { version = "0.30.6", extras = ["standard"] }
SQLAlchemy = "2.0.35"
alembic = "1.13.2"
pymysql = "1.1.1"
redis = "5.0.8"
celery = "5.4.0"
python-dotenv = "1.0.1"
loguru = "0.7.2"
pydantic-settings = "2.4.0"
python-jose = {version = "3.3.0", extras = ["cryptography"]}

[tool.poetry.group.dev.dependencies]
black = "24.8.0"
ruff = "0.6.8"
mypy = "1.11.2"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"


